+++ clients/fiat/src/rest-api/modules/fiat-api.ts
@@ -20,10 +20,8 @@ import {
     type RequestArgs,
 } from '@binance/common';
 import type {
-    FiatWithdrawResponse,
     GetFiatDepositWithdrawHistoryResponse,
     GetFiatPaymentsHistoryResponse,
-    GetOrderDetailResponse,
 } from '../types';
 
 /**
@@ -31,39 +29,6 @@ import type {
  */
 const FiatApiAxiosParamCreator = function (configuration: ConfigurationRestAPI) {
     return {
-        /**
-         * Submit withdraw request, in this version, we only support BRL withdrawal via bank_transfer.
-         *
-         * You need to call this api first, and call query order detail api in a loop to get the status of the order until this order is successful.
-         *
-         * Before calling this api, please make sure you have already completed your KYC or KYB, and already activated your fiat service on our website.
-         *
-         * you need to bind your bank account on web/app before using the corresponding account number
-         *
-         * Weight: 45000
-         *
-         * @summary Fiat Withdraw(WITHDRAW)
-         * @param {number | bigint} [recvWindow]
-         *
-         * @throws {RequiredError}
-         */
-        fiatWithdraw: async (recvWindow?: number | bigint): Promise<RequestArgs> => {
-            const localVarQueryParameter: Record<string, unknown> = {};
-
-            if (recvWindow !== undefined && recvWindow !== null) {
-                localVarQueryParameter['recvWindow'] = recvWindow;
-            }
-
-            let _timeUnit: TimeUnit | undefined;
-            if ('timeUnit' in configuration) _timeUnit = configuration.timeUnit as TimeUnit;
-
-            return {
-                endpoint: '/sapi/v2/fiat/withdraw',
-                method: 'GET',
-                params: localVarQueryParameter,
-                timeUnit: _timeUnit,
-            };
-        },
         /**
          * Get Fiat Deposit/Withdraw History
          *
@@ -197,46 +162,6 @@ const FiatApiAxiosParamCreator = function (configuration: ConfigurationRestAPI) 
                 timeUnit: _timeUnit,
             };
         },
-        /**
-         * Get Order Detail
-         *
-         * Before calling this api, please make sure you have already completed your KYC or KYB, and already activated your fiat service on our website.
-         *
-         * Weight: 45000
-         *
-         * @summary Get Order Detail(USER_DATA)
-         * @param {string} orderId order id retrieved from the api call of withdrawal
-         * @param {number | bigint} [recvWindow]
-         *
-         * @throws {RequiredError}
-         */
-        getOrderDetail: async (
-            orderId: string,
-            recvWindow?: number | bigint
-        ): Promise<RequestArgs> => {
-            // verify required parameter 'orderId' is not null or undefined
-            assertParamExists('getOrderDetail', 'orderId', orderId);
-
-            const localVarQueryParameter: Record<string, unknown> = {};
-
-            if (orderId !== undefined && orderId !== null) {
-                localVarQueryParameter['orderId'] = orderId;
-            }
-
-            if (recvWindow !== undefined && recvWindow !== null) {
-                localVarQueryParameter['recvWindow'] = recvWindow;
-            }
-
-            let _timeUnit: TimeUnit | undefined;
-            if ('timeUnit' in configuration) _timeUnit = configuration.timeUnit as TimeUnit;
-
-            return {
-                endpoint: '/sapi/v1/fiat/get-order-detail',
-                method: 'GET',
-                params: localVarQueryParameter,
-                timeUnit: _timeUnit,
-            };
-        },
     };
 };
 
@@ -245,26 +170,6 @@ const FiatApiAxiosParamCreator = function (configuration: ConfigurationRestAPI) 
  * @interface FiatApi
  */
 export interface FiatApiInterface {
-    /**
-     * Submit withdraw request, in this version, we only support BRL withdrawal via bank_transfer.
-     *
-     * You need to call this api first, and call query order detail api in a loop to get the status of the order until this order is successful.
-     *
-     * Before calling this api, please make sure you have already completed your KYC or KYB, and already activated your fiat service on our website.
-     *
-     * you need to bind your bank account on web/app before using the corresponding account number
-     *
-     * Weight: 45000
-     *
-     * @summary Fiat Withdraw(WITHDRAW)
-     * @param {FiatWithdrawRequest} requestParameters Request parameters.
-     *
-     * @throws {RequiredError | ConnectorClientError | UnauthorizedError | ForbiddenError | TooManyRequestsError | RateLimitBanError | ServerError | NotFoundError | NetworkError | BadRequestError}
-     * @memberof FiatApiInterface
-     */
-    fiatWithdraw(
-        requestParameters?: FiatWithdrawRequest
-    ): Promise<RestApiResponse<FiatWithdrawResponse>>;
     /**
      * Get Fiat Deposit/Withdraw History
      *
@@ -302,35 +207,6 @@ export interface FiatApiInterface {
     getFiatPaymentsHistory(
         requestParameters: GetFiatPaymentsHistoryRequest
     ): Promise<RestApiResponse<GetFiatPaymentsHistoryResponse>>;
-    /**
-     * Get Order Detail
-     *
-     * Before calling this api, please make sure you have already completed your KYC or KYB, and already activated your fiat service on our website.
-     *
-     * Weight: 45000
-     *
-     * @summary Get Order Detail(USER_DATA)
-     * @param {GetOrderDetailRequest} requestParameters Request parameters.
-     *
-     * @throws {RequiredError | ConnectorClientError | UnauthorizedError | ForbiddenError | TooManyRequestsError | RateLimitBanError | ServerError | NotFoundError | NetworkError | BadRequestError}
-     * @memberof FiatApiInterface
-     */
-    getOrderDetail(
-        requestParameters: GetOrderDetailRequest
-    ): Promise<RestApiResponse<GetOrderDetailResponse>>;
-}
-
-/**
- * Request parameters for fiatWithdraw operation in FiatApi.
- * @interface FiatWithdrawRequest
- */
-export interface FiatWithdrawRequest {
-    /**
-     *
-     * @type {number | bigint}
-     * @memberof FiatApiFiatWithdraw
-     */
-    readonly recvWindow?: number | bigint;
 }
 
 /**
@@ -430,26 +306,6 @@ export interface GetFiatPaymentsHistoryRequest {
 }
 
 /**
- * Request parameters for getOrderDetail operation in FiatApi.
- * @interface GetOrderDetailRequest
- */
-export interface GetOrderDetailRequest {
-    /**
-     * order id retrieved from the api call of withdrawal
-     * @type {string}
-     * @memberof FiatApiGetOrderDetail
-     */
-    readonly orderId: string;
-
-    /**
-     *
-     * @type {number | bigint}
-     * @memberof FiatApiGetOrderDetail
-     */
-    readonly recvWindow?: number | bigint;
-}
-
-/**
  * FiatApi - object-oriented interface
  * @class FiatApi
  */
@@ -460,40 +316,6 @@ export class FiatApi implements FiatApiInterface {
     constructor(configuration: ConfigurationRestAPI) {
         this.configuration = configuration;
         this.localVarAxiosParamCreator = FiatApiAxiosParamCreator(configuration);
-    }
-
-    /**
-     * Submit withdraw request, in this version, we only support BRL withdrawal via bank_transfer.
-     *
-     * You need to call this api first, and call query order detail api in a loop to get the status of the order until this order is successful.
-     *
-     * Before calling this api, please make sure you have already completed your KYC or KYB, and already activated your fiat service on our website.
-     *
-     * you need to bind your bank account on web/app before using the corresponding account number
-     *
-     * Weight: 45000
-     *
-     * @summary Fiat Withdraw(WITHDRAW)
-     * @param {FiatWithdrawRequest} requestParameters Request parameters.
-     * @returns {Promise<RestApiResponse<FiatWithdrawResponse>>}
-     * @throws {RequiredError | ConnectorClientError | UnauthorizedError | ForbiddenError | TooManyRequestsError | RateLimitBanError | ServerError | NotFoundError | NetworkError | BadRequestError}
-     * @memberof FiatApi
-     * @see {@link https://developers.binance.com/docs/fiat/rest-api/Fiat-Withdraw Binance API Documentation}
-     */
-    public async fiatWithdraw(
-        requestParameters: FiatWithdrawRequest = {}
-    ): Promise<RestApiResponse<FiatWithdrawResponse>> {
-        const localVarAxiosArgs = await this.localVarAxiosParamCreator.fiatWithdraw(
-            requestParameters?.recvWindow
-        );
-        return sendRequest<FiatWithdrawResponse>(
-            this.configuration,
-            localVarAxiosArgs.endpoint,
-            localVarAxiosArgs.method,
-            localVarAxiosArgs.params,
-            localVarAxiosArgs?.timeUnit,
-            { isSigned: true }
-        );
     }
 
     /**
@@ -563,37 +385,6 @@ export class FiatApi implements FiatApiInterface {
             requestParameters?.recvWindow
         );
         return sendRequest<GetFiatPaymentsHistoryResponse>(
-            this.configuration,
-            localVarAxiosArgs.endpoint,
-            localVarAxiosArgs.method,
-            localVarAxiosArgs.params,
-            localVarAxiosArgs?.timeUnit,
-            { isSigned: true }
-        );
-    }
-
-    /**
-     * Get Order Detail
-     *
-     * Before calling this api, please make sure you have already completed your KYC or KYB, and already activated your fiat service on our website.
-     *
-     * Weight: 45000
-     *
-     * @summary Get Order Detail(USER_DATA)
-     * @param {GetOrderDetailRequest} requestParameters Request parameters.
-     * @returns {Promise<RestApiResponse<GetOrderDetailResponse>>}
-     * @throws {RequiredError | ConnectorClientError | UnauthorizedError | ForbiddenError | TooManyRequestsError | RateLimitBanError | ServerError | NotFoundError | NetworkError | BadRequestError}
-     * @memberof FiatApi
-     * @see {@link https://developers.binance.com/docs/fiat/rest-api/Get-Order-Detail Binance API Documentation}
-     */
-    public async getOrderDetail(
-        requestParameters: GetOrderDetailRequest
-    ): Promise<RestApiResponse<GetOrderDetailResponse>> {
-        const localVarAxiosArgs = await this.localVarAxiosParamCreator.getOrderDetail(
-            requestParameters?.orderId,
-            requestParameters?.recvWindow
-        );
-        return sendRequest<GetOrderDetailResponse>(
             this.configuration,
             localVarAxiosArgs.endpoint,
             localVarAxiosArgs.method,